Reset session before loading a new project.
To do so, we save the DynFlags we had immediately after we started the GHC
session and save them in the Scion session state.  However, due to a bug in
GHC loading another project may still fail because GHC doesn't remove
instances when the package is not used anymore.

There are also some problems regarding unloading of interface files and
object code, but let's not worry about those, for now.