Implement frontend code for on-the-fly typechecking.
This needs to be made a little more robust, but it works so far.  Use

  M-x scion-turn-on-flycheck

to automatically typecheck tho current module whenever the file is saved or
the file has been edited (with a delay of 0.5 to 1 seconds).  However, the
latter does not actually save the file, so ATM it's a little pointless ;)