Fix build.  Sort of.
The mechanism to deal with transitive dependencies works only with
"cabal install" but not with the local directory "equivalents" for
"cabal configure" and "cabal build".

So we use "cabal install" to build the server which has the annoying
side effect of installing the binary globally (or per user).  Instead
of only in the current directory.  What a mess.