Use the Cabal library directly to configure projects.  Fixes #22.
So far we used GHCi to configure a project.  Unfortunately, that
becomes a problem when we have multiple versions of Cabal installed.
GHCi will simply pick the newest one.  We could add -hide-all-packages
and explicitly list which version of Cabal we want, but then Scion
needs to know which version of Cabal has been compiled with.

So instead we simply call Cabal directly to configure the project.
This has the positive side effect that Scion can now configure and
load Cabal itself (Issue 22).  Parsing extra args has become a little
more difficult now, though, so that's disabled for now.