Implement reading the PackageDB and the Names defined in it.
This is a prerequisite for automatically adding imports.  The `NameDB`
currently only maps names to the modules they are exported from.  In
the future we probably want types as well and possibly also source
location (where available) and where to find its documentation.

Generating this database is quite slow and memory hungry because it
indirectly causes GHC to load all the interfaces.  Generating is
therefore done offline and we just read the DB from a binary file.

For example, for my package DB consisting of 136 modules it requires
350 MB (185 MB resident) to generate the DB consisting of about 32K
names.  The Binary instance uses dictionaries to reduce the file size
which in my case is 2 MB.  Reading it back into memory takes about 16
MB (8.5 MB resident) and takes about half a second.

It would probably be worthwhile to investigate whether we can lazily
read this database.  (I think Hoogle does something like that.)